+++ NEWS.md
@@ -1,5 +1,7 @@
 # sigminer 1.1.0
 
+- Added `ms` strategy to select optimal solution by maximizing cosine similarity
+to reference signatures.
 - Added `same_size_clustering()` for same size clustering.
 - Added `show_cosmic()` to support reading COSMIC signatures in web browser (#288).
 - Changed argument `rel_threshold` behavior in `sig_fit()` and `get_sig_exposure()`.